functions/requirements.txt
functions-framework==3.*
google-api-python-client==2.*
google-auth==2.*
google-cloud-firestore==2.*
google-cloud-secret-manager==2.*
gitpython==3.*
pypandoc-binary==1.*
pdfplumber==0.*
markdownify==0.*
